Bicubic interpolator working correctly with tests.  When we initialize with just a function we use our old subsampling methodology inside each fitting cell.  When the user provides a function and its gradient however, we use the new method that only samples at the cells corners. This form should also ensure consistent gradients at interpolation cell boundaries.

+++ src/Utilities/XYInterpolatorInline.hh
@@ -84,10 +84,10 @@ void
 XYInterpolator::lowerBound(const double x, const double y,
                            size_t& ix, size_t& iy, size_t& i) const {
   ix = (mxlog ?
-        std::min(mnx1 - 1u, size_t(std::max(0.0, mnx1 + log((x - mAx)/mBx)))) :
+        std::min(mnx1 - 1u, size_t(std::max(0.0, double(mnx1) + log((x - mAx)/mBx)))) :
         std::min(mnx1 - 1u, size_t(std::max(0.0, x - mxmin)/mxstep)));
   iy = (mylog ?
-        std::min(mny1 - 1u, size_t(std::max(0.0, mny1 + log((y - mAy)/mBy)))) :
+        std::min(mny1 - 1u, size_t(std::max(0.0, double(mny1) + log((y - mAy)/mBy)))) :
         std::min(mny1 - 1u, size_t(std::max(0.0, y - mymin)/mystep)));
   CHECK(ix < mnx1 and iy < mny1);
   i = mncoeffs*(mnx1*iy + ix);
